# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cybersecurity Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BreachQueryType = typeof BreachQueryType[keyof typeof BreachQueryType];


export const BreachQueryType = {
  email: 'email',
  phone: 'phone',
  username: 'username',
  ip: 'ip',
  password: 'password',
  domain: 'domain',
} as const;

export interface BreachQuery {
  type: BreachQueryType;
  value: string;
}

export type BreachSourceRiskLevel = typeof BreachSourceRiskLevel[keyof typeof BreachSourceRiskLevel];


export const BreachSourceRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface BreachSource {
  name: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  date: string | null;
  /** @nullable */
  addedDate?: string | null;
  /** @nullable */
  domain?: string | null;
  dataClasses: string[];
  /** @nullable */
  pwnCount: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoPath?: string | null;
  isVerified?: boolean;
  isSensitive?: boolean;
  riskLevel: BreachSourceRiskLevel;
}

export type BreachResultQuery = {
  type: string;
  value: string;
};

export type BreachResultRiskLevel = typeof BreachResultRiskLevel[keyof typeof BreachResultRiskLevel];


export const BreachResultRiskLevel = {
  safe: 'safe',
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface BreachResult {
  found: boolean;
  query: BreachResultQuery;
  totalBreaches: number;
  /** @nullable */
  totalPwned?: number | null;
  /** 0-100 risk score */
  riskScore: number;
  riskLevel: BreachResultRiskLevel;
  sources: BreachSource[];
  tips: string[];
  /** @nullable */
  summary?: string | null;
}

export interface BreachEntry {
  name: string;
  title: string;
  domain: string;
  breachDate: string;
  addedDate: string;
  modifiedDate?: string;
  pwnCount: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoPath?: string | null;
  dataClasses: string[];
  isVerified: boolean;
  isFabricated?: boolean;
  isSensitive?: boolean;
  isRetired?: boolean;
  isSpamList?: boolean;
}

export type BreachStatsLargestBreach = {
  name: string;
  pwnCount: number;
};

export type BreachStatsNewestBreach = {
  name: string;
  date: string;
};

export type BreachStatsMostCommonDataTypesItem = {
  type: string;
  count: number;
};

export interface BreachStats {
  totalBreaches: number;
  totalPwnedAccounts: number;
  totalDataClasses: number;
  largestBreach: BreachStatsLargestBreach;
  newestBreach: BreachStatsNewestBreach;
  mostCommonDataTypes: BreachStatsMostCommonDataTypesItem[];
}