# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Cybersecurity Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /breach/check:
    post:
      operationId: checkBreach
      tags: [breach]
      summary: Check data for breaches
      description: Checks email, phone, username, IP, password, or domain against breach databases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreachQuery"
      responses:
        "200":
          description: Breach check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreachResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BreachQuery:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [email, phone, username, ip, password, domain]
        value:
          type: string
    BreachSource:
      type: object
      required: [name, date, dataClasses, pwnCount, riskLevel]
      properties:
        name:
          type: string
        title:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        addedDate:
          type: ["string", "null"]
        domain:
          type: ["string", "null"]
        dataClasses:
          type: array
          items:
            type: string
        pwnCount:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        logoPath:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isSensitive:
          type: boolean
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
    BreachResult:
      type: object
      required: [found, query, totalBreaches, riskScore, riskLevel, sources, tips]
      properties:
        found:
          type: boolean
        query:
          type: object
          required: [type, value]
          properties:
            type:
              type: string
            value:
              type: string
        totalBreaches:
          type: integer
        totalPwned:
          type: ["integer", "null"]
        riskScore:
          type: integer
          description: 0-100 risk score
        riskLevel:
          type: string
          enum: [safe, low, medium, high, critical]
        sources:
          type: array
          items:
            $ref: "#/components/schemas/BreachSource"
        tips:
          type: array
          items:
            type: string
        summary:
          type: ["string", "null"]
...