# lib/api-zod/src/generated/types/breachQuery.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cybersecurity Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { BreachQueryType } from './breachQueryType';

export interface BreachQuery {
  type: BreachQueryType;
  value: string;
}