# lib/api-zod/src/generated/types/breachResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cybersecurity Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { BreachResultQuery } from './breachResultQuery';
import type { BreachResultRiskLevel } from './breachResultRiskLevel';
import type { BreachSource } from './breachSource';

export interface BreachResult {
  found: boolean;
  query: BreachResultQuery;
  totalBreaches: number;
  /** @nullable */
  totalPwned?: number | null;
  /** 0-100 risk score */
  riskScore: number;
  riskLevel: BreachResultRiskLevel;
  sources: BreachSource[];
  tips: string[];
  /** @nullable */
  summary?: string | null;
}