# lib/api-zod/src/generated/types/breachSource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cybersecurity Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { BreachSourceRiskLevel } from './breachSourceRiskLevel';

export interface BreachSource {
  name: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  date: string | null;
  /** @nullable */
  addedDate?: string | null;
  /** @nullable */
  domain?: string | null;
  dataClasses: string[];
  /** @nullable */
  pwnCount: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoPath?: string | null;
  isVerified?: boolean;
  isSensitive?: boolean;
  riskLevel: BreachSourceRiskLevel;
}